Added chunked downloading and implemented it in the frontend; Organized storage stores a bit better; Updated the authentication to save the private key while session is active

+++ storage/src/data/app_file.rs
@@ -10,7 +10,7 @@ pub struct AppFile {
     pub encrypted_metadata: String,
     pub name_hash: String,
     pub mime: String,
-    pub size: Option<i32>,
+    pub size: Option<i64>,
     pub chunks: Option<i32>,
     pub chunks_stored: Option<i32>,
     pub file_id: Option<i32>,
